+++ src/reactpy_django/config.py
@@ -16,11 +16,10 @@ from reactpy_django.types import (
 )
 from reactpy_django.utils import import_dotted_path
 
-
 _logger = logging.getLogger(__name__)
 
 
-# Not user configurable settings
+# Non-configurable values
 REACTPY_DEBUG_MODE.set_current(getattr(settings, "DEBUG"))
 REACTPY_REGISTERED_COMPONENTS: dict[str, ComponentConstructor] = {}
 REACTPY_VIEW_COMPONENT_IFRAMES: dict[str, ViewComponentIframe] = {}
@@ -47,13 +46,16 @@ REACTPY_DATABASE: str = getattr(
     "REACTPY_DATABASE",
     DEFAULT_DB_ALIAS,
 )
+_default_query_postprocessor = getattr(
+    settings,
+    "REACTPY_DEFAULT_QUERY_POSTPROCESSOR",
+    None,
+)
 REACTPY_DEFAULT_QUERY_POSTPROCESSOR: AsyncPostprocessor | SyncPostprocessor | None = (
     import_dotted_path(
-        getattr(
-            settings,
-            "REACTPY_DEFAULT_QUERY_POSTPROCESSOR",
-            "reactpy_django.utils.django_query_postprocessor",
-        )
+        _default_query_postprocessor
+        if isinstance(_default_query_postprocessor, str)
+        else "reactpy_django.utils.django_query_postprocessor",
     )
 )
 REACTPY_AUTH_BACKEND: str | None = getattr(